adds basic switch casing for form, moved AvailableTimes component to form
